Actually cross-platformy.

+++ .zshrc
@@ -16,22 +16,70 @@ alias bower="noglob bower"
 alias npm="noglob npm"
 alias sass="noglob sass"
 
-alias fixopenwith="/System/Library/Frameworks/CoreServices.framework/Frameworks/LaunchServices.framework/Support/lsregister -kill -r -domain local -domain system -domain user"
-alias bup="brew update && brew upgrade"
 
-case 'uname' in
-    Darwin)
-        alias stink="say 'you stink'"
-        ;;
-    Linux)
-        alias update='sudo aptitude update'
-        alias upgrade='sudo aptitude upgrade'
-        ;;
-esac
+if [[ $('uname') == 'Darwin' ]]; then
+	
+	# Repair LaunchServices
+	alias fixopenwith="/System/Library/Frameworks/CoreServices.framework/Frameworks/LaunchServices.framework/Support/lsregister -kill -r -domain local -domain system -domain user"
+	
+	# brew upgrade
+	alias bup="brew update && brew upgrade"
 
-cdf() {
-    cd "`osascript -e 'tell app "Finder" to POSIX path of (insertion location as alias)'`"
-}
+	# cd to active Finder window
+	cdf() {
+		cd "`osascript -e 'tell app "Finder" to POSIX path of (insertion location as alias)'`"
+	}
+
+	# http://superuser.com/questions/313650/resume-zsh-terminal-os-x-lion/315029#answer-328148
+	# Tell the terminal about the working directory whenever it changes.
+	if [[ "$TERM_PROGRAM" == "Apple_Terminal" ]] && [[ -z "$INSIDE_EMACS" ]]; then
+
+	    update_terminal_cwd() {
+	        # Identify the directory using a "file:" scheme URL, including
+	        # the host name to disambiguate local vs. remote paths.
+
+	        # Percent-encode the pathname.
+	        local URL_PATH=''
+	        {
+	            # Use LANG=C to process text byte-by-byte.
+	            local i ch hexch LANG=C
+	            for ((i = 1; i <= ${#PWD}; ++i)); do
+	                ch="$PWD[i]"
+	                if [[ "$ch" =~ [/._~A-Za-z0-9-] ]]; then
+	                    URL_PATH+="$ch"
+	                else
+	                    hexch=$(printf "%02X" "'$ch")
+	                    URL_PATH+="%$hexch"
+	                fi
+	            done
+	        }
+
+	        local PWD_URL="file://$HOST$URL_PATH"
+	        #echo "$PWD_URL"        # testing
+	        printf '\e]7;%s\a' "$PWD_URL"
+	    }
+
+	    # Register the function so it is called whenever the working
+	    # directory changes.
+	    autoload add-zsh-hook
+	    add-zsh-hook chpwd update_terminal_cwd
+
+	    # Tell the terminal about the initial directory.
+	    update_terminal_cwd
+	fi
+
+elif [[ $('uname') == 'Linux' ]]; then
+	alias update='sudo apt-get update'
+	alias upgrade='sudo apt-get upgrade'
+
+fi
+
+
+# WP-CLI Bash completion
+autoload bashcompinit
+bashcompinit
+source $HOME/.wp-cli/vendor/wp-cli/wp-cli/utils/wp-completion.bash
+
 
 # Set to this to use case-sensitive completion
 # CASE_SENSITIVE="true"
@@ -53,50 +101,7 @@ COMPLETION_WAITING_DOTS="true"
 # Example format: plugins=(rails git textmate ruby lighthouse)
 plugins=(autojump bower emoji-clock git github gnu-utils history history-substring-search jump npm osx pip python sublime terminalapp vagrant zsh-syntax-highlighting)
 
-# http://superuser.com/questions/313650/resume-zsh-terminal-os-x-lion/315029#answer-328148
-# Tell the terminal about the working directory whenever it changes.
-if [[ "$TERM_PROGRAM" == "Apple_Terminal" ]] && [[ -z "$INSIDE_EMACS" ]]; then
-
-    update_terminal_cwd() {
-        # Identify the directory using a "file:" scheme URL, including
-        # the host name to disambiguate local vs. remote paths.
-
-        # Percent-encode the pathname.
-        local URL_PATH=''
-        {
-            # Use LANG=C to process text byte-by-byte.
-            local i ch hexch LANG=C
-            for ((i = 1; i <= ${#PWD}; ++i)); do
-                ch="$PWD[i]"
-                if [[ "$ch" =~ [/._~A-Za-z0-9-] ]]; then
-                    URL_PATH+="$ch"
-                else
-                    hexch=$(printf "%02X" "'$ch")
-                    URL_PATH+="%$hexch"
-                fi
-            done
-        }
-
-        local PWD_URL="file://$HOST$URL_PATH"
-        #echo "$PWD_URL"        # testing
-        printf '\e]7;%s\a' "$PWD_URL"
-    }
-
-    # Register the function so it is called whenever the working
-    # directory changes.
-    autoload add-zsh-hook
-    add-zsh-hook chpwd update_terminal_cwd
-
-    # Tell the terminal about the initial directory.
-    update_terminal_cwd
-fi
-
 source $ZSH/oh-my-zsh.sh
-
-# WP-CLI Bash completion
-autoload bashcompinit
-bashcompinit
-source $HOME/.wp-cli/vendor/wp-cli/wp-cli/utils/wp-completion.bash
 
 export CODAPATH="/Users/nateb/Applications/Coda\ 2.app:"
 export SVN_EDITOR="subl"
